Changes for PR on restructuring of b2/core 1. endgames/events.hpp now does not have commented out tracking/events.hpp code. A few commented prototypes for Bertini event types are left for future development of events in the endgames.
2. In endgames/observers.hpp the file has been reduced to two namespaces and no observers have been implemented.

3. Removal of core/bertini2, core/generating_test, core/nag_algorithms_test, and python/ltmain.sh

This commit should address issues for the PR for restructuring of the b2/core.
